requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
numpy
pandas
scikit-learn
joblib
python-dotenv
authlib
